requirements.txt
aioredis==1.3.1
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.6
certifi==2022.6.15
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.1.0
cheroot==8.6.0
CherryPy==18.6.1
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
drf-flex-fields==0.9.8
drf-yasg==1.21.3
gevent==21.12.0
greenlet==1.1.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.7.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
more-itertools==8.13.0
msgpack==1.0.3
packaging==21.3
portend==3.1.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2022.1
PyYAML==6.0
redis==4.2.2
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tempora==5.0.1
toml==0.10.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.2
wrapt==1.14.0
zc.lockfile==2.0
zope.event==4.5.0
zope.interface==5.4.0
